Extract DirectoryOptions internal structure to core package
